// ai-docs/src/52_graphql/fixtures/github/issues.graphql.ts
export class IssueSummary extends Schema.Opaque<IssueSummary>()(Schema.Struct({
  id: Shared.ID,
  /** Identifies the issue number. */
  number: Shared.Int,
  /** Identifies the issue title. */
  title: Schema.String,
  /** Identifies the state of the issue. */
  state: GraphQL.enumLiterals(Shared.IssueState.literals),
  /** The HTTP URL for this issue */
  url: Shared.URI,
  /** Identifies the date and time when the object was created. */
  createdAt: Shared.DateTime
})) {}

/** `query RepoIssues` */
export const RepoIssues = GraphQL.query("RepoIssues", {
  document: "query RepoIssues($owner:String!$name:String!$states:[IssueState!]$after:String){repository(owner:$owner name:$name){issues(first:50 after:$after states:$states){pageInfo{hasNextPage endCursor}nodes{...IssueSummary}}}}fragment IssueSummary on Issue{id number title state url createdAt}",
  variables: {
    owner: Schema.String,
    name: Schema.String,
    states: Schema.optional(Schema.NullOr(Schema.Array(Shared.IssueState))),
    after: Schema.optional(Schema.NullOr(Schema.String))
  },
  result: Schema.Struct({
    /** Lookup a given repository by the owner and repository name. */
    repository: Schema.NullOr(Schema.Struct({
      /** A list of issues that have been opened in the repository. */
      issues: Schema.Struct({
        /** Information to aid in pagination. */
        pageInfo: Schema.Struct({
          /** When paginating forwards, are there more items? */
          hasNextPage: Schema.Boolean,
          /** When paginating forwards, the cursor to continue. */
          endCursor: Schema.NullOr(Schema.String)
        }),
        /** A list of nodes. */
        nodes: Schema.NullOr(Schema.Array(Schema.NullOr(IssueSummary)))
      })
    }))
  })
})
export declare namespace RepoIssues {
  export type Variables = typeof RepoIssues.variables.Type
  export type Result = typeof RepoIssues.result.Type
}

/** `query RepoIssueEdges` */
export const RepoIssueEdges = GraphQL.query("RepoIssueEdges", {
  document: "query RepoIssueEdges($owner:String!$name:String!$after:String){repository(owner:$owner name:$name){issues(first:50 after:$after){pageInfo{hasNextPage endCursor}edges{cursor node{...IssueSummary}}}}}fragment IssueSummary on Issue{id number title state url createdAt}",
  variables: {
    owner: Schema.String,
    name: Schema.String,
    after: Schema.optional(Schema.NullOr(Schema.String))
  },
  result: Schema.Struct({
    /** Lookup a given repository by the owner and repository name. */
    repository: Schema.NullOr(Schema.Struct({
      /** A list of issues that have been opened in the repository. */
      issues: Schema.Struct({
        /** Information to aid in pagination. */
        pageInfo: Schema.Struct({
          /** When paginating forwards, are there more items? */
          hasNextPage: Schema.Boolean,
          /** When paginating forwards, the cursor to continue. */
          endCursor: Schema.NullOr(Schema.String)
        }),
        /** A list of edges. */
        edges: Schema.NullOr(Schema.Array(Schema.NullOr(Schema.Struct({
          /** A cursor for use in pagination. */
          cursor: Schema.String,
          /** The item at the end of the edge. */
          node: Schema.NullOr(IssueSummary)
        }))))
      })
    }))
  })
})
export declare namespace RepoIssueEdges {
  export type Variables = typeof RepoIssueEdges.variables.Type
  export type Result = typeof RepoIssueEdges.result.Type
}

/** `mutation AddComment` */
export const AddComment = GraphQL.mutation("AddComment", {
  document: "mutation AddComment($subjectId:ID!$body:String!){addComment(input:{subjectId:$subjectId body:$body}){clientMutationId}}",
  variables: {
    subjectId: Shared.ID,
    body: Schema.String
  },
  result: Schema.Struct({
    /** Adds a comment to an Issue or Pull Request. */
    addComment: Schema.NullOr(Schema.Struct({
      /** A unique identifier for the client performing the mutation. */
      clientMutationId: Schema.NullOr(Schema.String)
    }))
  })
})
export declare namespace AddComment {
  export type Variables = typeof AddComment.variables.Type
  export type Result = typeof AddComment.result.Type
}

/** Every operation in issues.graphql. */
export const IssuesGroup = GraphQLGroup.make(RepoIssues, RepoIssueEdges, AddComment)