// ai-docs/src/52_graphql/fixtures/github/pullRequests.graphql.ts
export const OpenPullRequestIds = GraphQL.query("OpenPullRequestIds", {
  document: "query OpenPullRequestIds($owner:String!$name:String!$after:String){repository(owner:$owner name:$name){pullRequests(first:100 after:$after states:[OPEN]){pageInfo{hasNextPage endCursor}nodes{id}}}}",
  variables: {
    owner: Schema.String,
    name: Schema.String,
    after: Schema.optional(Schema.NullOr(Schema.String))
  },
  result: Schema.Struct({
    /** Lookup a given repository by the owner and repository name. */
    repository: Schema.NullOr(Schema.Struct({
      /** A list of pull requests that have been opened in the repository. */
      pullRequests: Schema.Struct({
        /** Information to aid in pagination. */
        pageInfo: Schema.Struct({
          /** When paginating forwards, are there more items? */
          hasNextPage: Schema.Boolean,
          /** When paginating forwards, the cursor to continue. */
          endCursor: Schema.NullOr(Schema.String)
        }),
        /** A list of nodes. */
        nodes: Schema.NullOr(Schema.Array(Schema.NullOr(Schema.Struct({
          id: Shared.ID
        }))))
      })
    }))
  })
})
export declare namespace OpenPullRequestIds {
  export type Variables = typeof OpenPullRequestIds.variables.Type
  export type Result = typeof OpenPullRequestIds.result.Type
}

/** `query PullRequestsById` */
export const PullRequestsById = GraphQL.query("PullRequestsById", {
  document: "query PullRequestsById($ids:[ID!]!){nodes(ids:$ids){__typename...on PullRequest{id number title merged}}}",
  variables: {
    ids: Schema.Array(Shared.ID)
  },
  result: Schema.Struct({
    /** Lookup nodes by a list of IDs. */
    nodes: Schema.Array(Schema.NullOr(Schema.Union([
      Schema.Struct({
        __typename: Schema.Literal("PullRequest"),
        id: Shared.ID,
        /** Identifies the pull request number. */
        number: Shared.Int,
        /** Identifies the pull request title. */
        title: Schema.String,
        /** Whether or not the pull request was merged. */
        merged: Schema.Boolean
      }),
      Schema.Struct({
        __typename: GraphQL.otherTypename<Shared.Typename.Node>()(["PullRequest"])
      })
    ])))
  })
})
export declare namespace PullRequestsById {
  export type Variables = typeof PullRequestsById.variables.Type
  export type Result = typeof PullRequestsById.result.Type
}

/** `query EffectRepositories` */
export const EffectRepositories = GraphQL.query("EffectRepositories", {
  document: "query EffectRepositories{effect:repository(owner:\"Effect-TS\" name:\"effect\"){nameWithOwner}website:repository(owner:\"Effect-TS\" name:\"website\"){nameWithOwner}rateLimit{cost remaining}}",
  result: Schema.Struct({
    /** Lookup a given repository by the owner and repository name. */
    effect: Schema.NullOr(Schema.Struct({
      /** The repository's name with owner. */
      nameWithOwner: Schema.String
    })),
    /** Lookup a given repository by the owner and repository name. */
    website: Schema.NullOr(Schema.Struct({
      /** The repository's name with owner. */
      nameWithOwner: Schema.String
    })),
    /** The client's rate limit information. */
    rateLimit: Schema.NullOr(Schema.Struct({
      /** The point cost for the current query counting against the rate limit. */
      cost: Shared.Int,
      /** The number of points remaining in the current rate limit window. */
      remaining: Shared.Int
    }))
  })
})
export declare namespace EffectRepositories {
  export type Variables = typeof EffectRepositories.variables.Type
  export type Result = typeof EffectRepositories.result.Type
}

/** Every operation in pullRequests.graphql. */
export const PullRequestsGroup = GraphQLGroup.make(OpenPullRequestIds, PullRequestsById, EffectRepositories)